Fixed minor bug with theme swap and restructured components folder.
