+++ src/components/flamelink/FlameLinkCollectionGalleryContent.js
@@ -1,6 +1,6 @@
 import { Component } from 'react';
-import flamelinkApp from '../flamelink.js';
-import '../css/FlameLink.css';
+import flamelinkApp from '../../flamelink.js';
+import '../../css/FlameLink.css';
 
 class FlameLinkCollectionGalleryContent extends Component {
     constructor(props) {
