+++ src/components/flamelink/FlameLinkCollectionStructure.js
@@ -2,7 +2,7 @@ import React, { Component} from 'react';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
-import flamelinkApp from '../flamelink.js';
+import flamelinkApp from '../../flamelink.js';
 import FlameLinkImage from './FlameLinkImage';
 import FlameLinkFieldSet from './FlameLinkFieldSet';
 
@@ -14,7 +14,7 @@ const styles = theme => ({
         },
     });
 
-class FlameLinkStructure extends Component {
+class FlameLinkCollectionStructure extends Component {
     constructor() {
         super();
         
@@ -30,54 +30,54 @@ class FlameLinkStructure extends Component {
         }))
       }
 
-    getContent(schemaField, key, type, description){
-            if (type === 'text'){
+    getContent(content, field, key, type, description){
+        if (type === 'text'){
                 if(description === 'h1'){
                     return  (
                             <Typography variant='display4' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 if(description === 'h2'){
                     return  (
                             <Typography variant='display3' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 if(description === 'h3'){
                     return  (
                             <Typography variant='display2' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 if(description === 'h4'){
                     return  (
                             <Typography variant='display1' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 if(description === 'h5'){
                     return  (
                             <Typography variant='headline' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 if(description === 'h6'){
                     return  (
                             <Typography variant='title' id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )                
                 }
                 else{
                     return  (
                             <Typography variant='body2' component="p" id={key}>
-                                {this.state.schemaContent[key]}
+                                {content[key]}
                             </Typography>
                             )
                 }
@@ -85,22 +85,22 @@ class FlameLinkStructure extends Component {
             if(type === 'textarea'){
                 return  (
                         <Typography variant='body2' component="p" id={key}>
-                            {this.state.schemaContent[key]}
+                            {content[key]}
                         </Typography>
                         )               
             }
             if (type === 'media'){
-                for (var val in this.state.schemaContent[key]){
-                    global.mediaID = this.state.schemaContent[key][val];
+                for (var val in content[key]){
+                    global.mediaID = content[key][val];
                     return <FlameLinkImage/>
                 } 
             }
             if (type === 'fieldset'){
-                if(this.state.schemaContent === ''){
+                if(content === ''){
                     return
                 }
                 else{
-                    return <FlameLinkFieldSet field={this.state.schemaContent[key]} field2={schemaField.options}/>
+                    return <FlameLinkFieldSet field={content[key]} field2={field.options}/>
                 }
             }
     }
@@ -114,10 +114,10 @@ class FlameLinkStructure extends Component {
         const xs = this.props.field.gridColumns.xs;
         return(
             <Grid item lg={lg} md={md} sm={sm} xs={xs} className={classes.flamelinkItem}>
-                    {this.getContent(this.props.field, this.props.field.key, this.props.type, this.props.field.description)}
+                    {this.getContent(this.props.schemaContent, this.props.field, this.props.field.key, this.props.type, this.props.field.description)}
             </Grid>
         );
     }
 }
 
-export default withStyles(styles)(FlameLinkStructure);
+export default withStyles(styles)(FlameLinkCollectionStructure);
