+++ src/components/flamelink/FlameLinkImage.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Typography from '@material-ui/core/Typography';
-import flamelinkApp from '../flamelink.js';
-import '../css/FlameLink.css';
+import flamelinkApp from '../../flamelink.js';
+import '../../css/FlameLink.css';
 
 class FlameLinkImage extends Component {
     constructor() {
