+++ src/components/list/SightingDetail.js
