+++ src/pages/Info.js
@@ -1,5 +1,5 @@
 import React, { Component, Fragment } from 'react';
-import FlameLinkComponentCreations from '../components/FlameLinkComponentCreations';
+import FlameLinkComponentCreations from '../components/flamelink/FlameLinkComponentCreations';
 import flamelinkApp from '../flamelink.js';
 
 class Info extends Component {
