+++ src/pages/Report.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import ReportForm from '../components/ReportForm';
+import ReportForm from '../components/forms/ReportForm';
 
 class Report extends Component {
     componentDidMount() {
